requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
langchain==0.2.17
langchain-core==0.2.43
langchain-community==0.2.16
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
faiss-cpu==1.9.0.post1
tiktoken==0.7.0
